pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bgp-harvest"
version = "0.1.0"
description = "BGP route harvesting and validation pipeline"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6.0",
  "clickhouse-driver>=0.2.0",
  "clickhouse-connect>=0.7.0",
  "pybgpstream>=2.0.2",
  "requests==2.32.4",
  "tqdm>=4.66.2",
  "urllib3==1.26.6",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.11.0",
  "pytest>=8.3.0",
  "ruff>=0.6.0",
]

[project.scripts]
bgp-harvest = "bgp_harvest.cli.main:main"

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
bgp_harvest = ["clients/sql/*.sql"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.mypy]
python_version = "3.11"
packages = ["bgp_harvest"]
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
strict_optional = true
disallow_untyped_defs = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = ["yaml", "requests", "requests.adapters", "pybgpstream", "clickhouse_driver", "clickhouse_connect"]
ignore_missing_imports = true
